More intuitive meeaning of `Split` `direction`
As the writer of the Split component, I myself is still confused between
`vertical` or `horizontal` (is it the about the split line or how the
children are laid out). Most of the time, I thought about the split
line. So I think it's bettet to change the meaning of `direction` to
reflex that.